requirements.txt
Flask==2.2.2
Pillow==9.4.0
Pillow==9.0.1
Pillow==10.2.0
selenium==4.17.2
